Improved Resource ownership (#39)
Fixes #35 
Fixes #36 

The primary issue that was causing #36 is shared_ptr creep. Usually the deferred deletion / pseudo-garbage collection enabled by shared_ptrs is useful, but in this case, the resources that own the worker threads are held in objects that are managed by shared_ptrs. In some cases, the worker threads are the last owners of these shared ptrs and thus exhibit a deadlock where a thread ultimately tries to join itself.

This PR removes all shared_ptrs from object/resources that could be potentially captured by worker threads and replaces them with references.

The code is structured so that all resources and worker threads are owned by the `Executor` -> `resources::Manager` and that all threads must be joined before `resources::Manager` is destroyed. 

This ensures that the resources outlive the pipeline execution provided that the Executor stays in scope. If it goes out of scope or is destroyed it should issue a kill to the pipeline, but still block until all threads are joined.

All executor resources should be created and managed by `resources::Manager` as a single point for life cycle management.

Authors:
  - Ryan Olson (https://github.com/ryanolson)
  - Michael Demoret (https://github.com/mdemoret-nv)

Approvers:
  - Michael Demoret (https://github.com/mdemoret-nv)

URL: https://github.com/nv-morpheus/SRF/pull/39