fix faulty assumption about remote key sizes (#137)
I had put an assumption that the memory required to store UCX remote keys for GPU memory would always be larger than for Host memory.

This assumption is incorrect. The size of buffer for the keys is dependent on which transports are enable, the number of NICs, the number of GPUs and a variety of other runtime variables used to initialize UCX.

I've left a comment in the tests to remind developers and users of that fact.

Authors:
  - Ryan Olson (https://github.com/ryanolson)

Approvers:
  - Michael Demoret (https://github.com/mdemoret-nv)

URL: https://github.com/nv-morpheus/SRF/pull/137