Update PortBuilder to Work with Types That Do Not Have a Default Constructor (#165)
If you create a source or a sink class that derives from `PythonSink`, `PythonSource` or `PythonNode` and the message type does not have a default constructor, you will run into a compilation error. This was due to the fact that the new Port auto registration added in PR #153 tries to compile ingress/egress ports for both `T` and `shared_ptr<T>`.

This PR adds a check to skip compiling a port with `T` if `T` does not have a default constructor.

Authors:
  - Michael Demoret (https://github.com/mdemoret-nv)

Approvers:
  - Devin Robison (https://github.com/drobison00)

URL: https://github.com/nv-morpheus/SRF/pull/165