Update `sccache` bucket
This PR updates the `sccache` configuration settings to use a new bucket, `rapids-sccache-east`.

Unlike the previous `rapids-sccache` bucket, `rapids-sccache-east` resides in the same AWS region as the rest of our CI infrastructure (`us-east-2`).

This should result in faster, more reliable `sccache` connections and will also help keep our data transfer costs down.

**Important Note**: The changes from [this `sccache` PR](https://github.com/mozilla/sccache/pull/1403) are required to use a bucket in `us-east-2`. These changes were incorporated in `sccache` `v0.3.2`, so you'll need that version or later (preferably the latest, `v0.3.3`). Please ensure that your CI images satisfy this constraint before merging this PR.

Additionally, this PR updates the `SCCACHE_REGION` to the region of the bucket. This value is not meant to be a variable.