Dynamically loading `libnvidia-ml.so.1` instead of directly linking (#313)
Currently, we install the driver into one of the CI images to allow for stub generation during compilation. This was needed because `device_info.cpp` linked directly against `CUDA::nvml`. This caused a link dependency on a driver library which is problematic when building with CPU-only docker images.

Instead, we dynamically load `libnvidia-ml.so.1` (appending the `.so.1` to avoid collisions with the stub file `libnvidia-ml.so`) and the necessary functions at runtime. If the library is not found, using a GPU will be disabled. This allows loading of the library for stub generation without needing a GPU.

Authors:
  - Michael Demoret (https://github.com/mdemoret-nv)

Approvers:
  - David Gardner (https://github.com/dagardner-nv)
  - Devin Robison (https://github.com/drobison00)

URL: https://github.com/nv-morpheus/MRC/pull/313